# contracts/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API for the Reverse Marketplace platform",
    "title": "Reverse Marketplace API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/admin/users/{userId}": {
      "get": {
        "description": "Get user details",
        "parameters": [
          {
            "description": "User UUID",
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/v1/sellers/user/{userId}": {
      "get": {
        "description": "Get seller profile by user ID",
        "parameters": [
          {
            "description": "User UUID",
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "Sellers"
        ]
      }
    },
...
    "/api/v1/users/me": {
      "delete": {
        "description": "Delete account (requires password confirmation)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "password": {
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "password"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ]
      },
      "get": {
        "description": "Get current user profile",
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "description": "Update current user profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "bio": {
                    "description": "Short bio about yourself",
                    "maxLength": 2000,
                    "nullable": true,
                    "type": "string"
                  },
                  "firstName": {
                    "description": "First name (e.g. John)",
                    "maxLength": 100,
                    "minLength": 1,
                    "type": "string"
                  },
                  "lastName": {
                    "description": "Last name (e.g. Doe)",
                    "maxLength": 100,
                    "minLength": 1,
                    "type": "string"
                  },
                  "latitude": {
                    "description": "Latitude (e.g. 32.7767)",
                    "maximum": 90,
                    "minimum": -90,
                    "nullable": true,
                    "type": "number"
                  },
                  "locationCity": {
                    "description": "City (e.g. Dallas)",
                    "maxLength": 100,
                    "nullable": true,
                    "type": "string"
                  },
                  "locationCountry": {
                    "description": "Country (default: US)",
                    "maxLength": 50,
                    "type": "string"
                  },
                  "locationState": {
                    "description": "State (e.g. TX)",
                    "maxLength": 50,
                    "nullable": true,
                    "type": "string"
                  },
                  "locationZip": {
                    "description": "ZIP code (e.g. 75001)",
                    "nullable": true,
                    "pattern": "^\\d{5}(-\\d{4})?$",
                    "type": "string"
                  },
                  "longitude": {
                    "description": "Longitude (e.g. -96.7970)",
                    "maximum": 180,
                    "minimum": -180,
                    "nullable": true,
                    "type": "number"
                  },
                  "notificationPreferences": {
                    "description": "Notification preferences",
                    "properties": {
                      "email_offers": {
                        "description": "Email notifications for offers",
                        "type": "boolean"
                      },
                      "push_messages": {
                        "description": "Push notifications for messages",
                        "type": "boolean"
                      },
                      "sms_offers": {
                        "description": "SMS notifications for offers",
                        "type": "boolean"
                      }
                    },
                    "type": "object"
                  },
                  "phone": {
                    "description": "Phone number (e.g. 214-555-0100)",
                    "maxLength": 20,
                    "nullable": true,
                    "type": "string"
                  },
                  "preferredLanguage": {
                    "description": "Preferred language code (e.g. en, es, zh)",
                    "enum": [
                      "en",
                      "es",
                      "zh",
                      "ar",
                      "fr",
                      "pt",
                      "hi",
                      "vi",
                      "ko",
                      "ja"
                    ],
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me/account-type": {
      "patch": {
        "description": "Switch account type",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "accountType": {
                    "description": "New account type",
                    "enum": [
                      "buyer",
                      "seller",
                      "both"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "accountType"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me/business-profile": {
      "patch": {
        "description": "Attach or update the sales-tax certificate (and optionally name/type/EIN) for an existing business account — used to complete business registration after the upload endpoint becomes reachable (auth in hand)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "businessName": {
                    "description": "Legal business name — optional, only if correcting",
                    "maxLength": 255,
                    "minLength": 1,
                    "type": "string"
                  },
                  "businessType": {
                    "description": "Business entity type — optional, only if correcting",
                    "enum": [
                      "llc",
                      "corporation",
                      "sole_proprietor",
                      "partnership"
                    ],
                    "type": "string"
                  },
                  "ein": {
                    "description": "Employer Identification Number (XX-XXXXXXX) — optional, only if correcting",
                    "pattern": "^\\d{2}-\\d{7}$",
                    "type": "string"
                  },
                  "salesTaxCertificateUrl": {
                    "description": "R2 URL of uploaded sales-tax certificate",
                    "format": "uri",
                    "type": "string"
                  }
                },
                "required": [
                  "salesTaxCertificateUrl"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me/change-password": {
      "post": {
        "description": "Change password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "currentPassword": {
                    "description": "Your current password",
                    "minLength": 1,
                    "type": "string"
                  },
                  "newPassword": {
                    "allOf": [
                      {
                        "pattern": "[A-Z]",
                        "type": "string"
                      },
                      {
                        "pattern": "[a-z]",
                        "type": "string"
                      },
                      {
                        "pattern": "[0-9]",
                        "type": "string"
                      },
                      {
                        "pattern": "[^A-Za-z0-9]",
                        "type": "string"
                      }
                    ],
                    "description": "New password (e.g. NewSecure456!)",
                    "minLength": 8,
                    "type": "string"
                  }
                },
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me/fcm-token": {
      "put": {
        "description": "Update FCM push token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "fcmToken": {
                    "description": "Firebase Cloud Messaging device token",
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "fcmToken"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me/marketplace-context": {
      "put": {
        "description": "Switch active marketplace context (B2C, B2B, C2C)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "context": {
                    "description": "Marketplace context to switch to (b2c, b2b, c2c)",
                    "enum": [
                      "b2c",
                      "b2b",
                      "c2c"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "context"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me/photo": {
      "patch": {
        "description": "Update profile photo URL",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "photoUrl": {
                    "format": "uri",
                    "type": "string"
                  }
                },
                "required": [
                  "photoUrl"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/me/upgrade-to-business": {
      "post": {
        "description": "Upgrade classic account to business — adds EIN, business name/type, sales-tax certificate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "businessName": {
                    "description": "Legal business name",
                    "maxLength": 255,
                    "minLength": 1,
                    "type": "string"
                  },
                  "businessType": {
                    "description": "Business entity type",
                    "enum": [
                      "llc",
                      "corporation",
                      "sole_proprietor",
                      "partnership"
                    ],
                    "type": "string"
                  },
                  "ein": {
                    "description": "Employer Identification Number (XX-XXXXXXX)",
                    "pattern": "^\\d{2}-\\d{7}$",
                    "type": "string"
                  },
                  "salesTaxCertificateUrl": {
                    "description": "R2 URL of uploaded sales-tax certificate",
                    "format": "uri",
                    "type": "string"
                  }
                },
                "required": [
                  "ein",
                  "businessName",
                  "businessType",
                  "salesTaxCertificateUrl"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{userId}": {
      "get": {
        "description": "Get public user profile",
        "parameters": [
          {
            "description": "User UUID",
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Reverse Marketplace API (base path)",
      "url": "/api/v1"
    }
  ]
}